Move Gdi32 P/Invoke to Interop.*.cs
