+++ src/libraries/Common/src/Interop/Windows/Gdi32/Interop.EndDoc.cs
@@ -1,0 +1,20 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System.Runtime.InteropServices;
+#if NET7_0_OR_GREATER
+using System.Runtime.InteropServices.Marshalling;
+#endif
+
+internal static partial class Interop
+{
+    internal static partial class Gdi32
+    {
+        [LibraryImport(Libraries.Gdi32, SetLastError = true)]
+        internal static partial int EndDoc(
+#if NET7_0_OR_GREATER
+            [MarshalUsing(typeof(HandleRefMarshaller))]
+#endif
+            HandleRef hDC);
+    }
+}
