+++ src/libraries/Common/src/Interop/Windows/Gdi32/Interop.ExtEscape.cs
@@ -1,0 +1,31 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System.Runtime.InteropServices;
+#if NET7_0_OR_GREATER
+using System.Runtime.InteropServices.Marshalling;
+#endif
+
+internal static partial class Interop
+{
+    internal static partial class Gdi32
+    {
+        internal const int QUERYESCSUPPORT = 8;
+        internal const int CHECKJPEGFORMAT = 4119;
+        internal const int CHECKPNGFORMAT = 4120;
+
+        [LibraryImport(Libraries.Gdi32, SetLastError = true)]
+        internal static partial int ExtEscape(
+#if NET7_0_OR_GREATER
+            [MarshalUsing(typeof(HandleRefMarshaller))]
+#endif
+            HandleRef hDC, int nEscape, int cbInput, ref int inData, int cbOutput, out int outData);
+
+        [LibraryImport(Libraries.Gdi32, SetLastError = true)]
+        internal static partial int ExtEscape(
+#if NET7_0_OR_GREATER
+            [MarshalUsing(typeof(HandleRefMarshaller))]
+#endif
+            HandleRef hDC, int nEscape, int cbInput, byte[] inData, int cbOutput, out int outData);
+    }
+}
