+++ src/libraries/Common/src/Interop/Windows/Gdi32/Interop.GetPaletteEntries.cs
@@ -1,0 +1,20 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System.Runtime.InteropServices;
+#if NET7_0_OR_GREATER
+using System.Runtime.InteropServices.Marshalling;
+#endif
+
+internal static partial class Interop
+{
+    internal static partial class Gdi32
+    {
+        [LibraryImport(Libraries.Gdi32)]
+        internal static partial uint GetPaletteEntries(
+#if NET7_0_OR_GREATER
+            [MarshalUsing(typeof(HandleRefMarshaller))]
+#endif
+            HandleRef hpal, int iStartIndex, int nEntries, byte[] lppe);
+    }
+}
