+++ src/libraries/System.Drawing.Common/src/System.Drawing.Common.csproj
@@ -214,7 +214,6 @@ Since .NET 7, non-Windows platforms are not supported, even with the runtime con
     <Compile Include="misc\GDI\DeviceContextType.cs" />
     <Compile Include="misc\GDI\WindowsGraphics.cs" />
     <Compile Include="misc\GDI\WindowsRegion.cs" />
-    <Compile Include="Interop\Windows\Interop.Gdi32.cs" />
 
     <Compile Include="$(CommonPath)Interop\Windows\Interop.BOOL.cs"
              Link="Common\Interop\Windows\Interop.BOOL.cs" />
@@ -222,12 +221,22 @@ Since .NET 7, non-Windows platforms are not supported, even with the runtime con
              Link="Common\Interop\Windows\Interop.Libraries.cs" />
     <Compile Include="$(CommonPath)Interop\Windows\Comdlg32\Interop.PrintDlg.cs"
              Link="Common\Interop\Windows\Comdlg32\Interop.PrintDlg.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.AbortDoc.cs"
+             Link="Common\Interop\Windows\Gdi32\Interop.AbortDoc.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.AddFontResourceEx.cs"
+             Link="Common\Interop\Windows\Gdi32\Interop.AddFontResourceEx.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.BITMAPINFO_FLAT.cs"
+             Link="Common\Interop\Windows\Gdi32\Interop.BITMAPINFO_FLAT.cs" />
     <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.CombineRgn.cs"
              Link="Common\Interop\Windows\Gdi32\Interop.CombineRgn.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.CreateCompatibleBitmap.cs"
+             Link="Common\Interop\Windows\Gdi32\Interop.CreateCompatibleBitmap.cs" />
     <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.CreateCompatibleDC.cs"
              Link="Common\Interop\Windows\Gdi32\Interop.CreateCompatibleDC.cs" />
     <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.CreateDC.cs"
              Link="Common\Interop\Windows\Gdi32\Interop.CreateDC.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.CreateDIBSection.cs"
+             Link="Common\Interop\Windows\Gdi32\Interop.CreateDIBSection.cs" />
     <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.CreateIC.cs"
              Link="Common\Interop\Windows\Gdi32\Interop.CreateIC.cs" />
     <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.CreateFontIndirect.cs"
@@ -238,18 +247,40 @@ Since .NET 7, non-Windows platforms are not supported, even with the runtime con
              Link="Common\Interop\Windows\Gdi32\Interop.DeleteDC.cs" />
     <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.DeleteObject.cs"
              Link="Common\Interop\Windows\Gdi32\Interop.DeleteObject.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.DEVMODE.cs"
+             Link="Common\Interop\Windows\Gdi32\Interop.DEVMODE.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.EndDoc.cs"
+             Link="Common\Interop\Windows\Gdi32\Interop.EndDoc.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.EndPage.cs"
+             Link="Common\Interop\Windows\Gdi32\Interop.EndPage.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.ExtEscape.cs"
+             Link="Common\Interop\Windows\Gdi32\Interop.ExtEscape.cs" />
     <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.GetClipRgn.cs"
              Link="Common\Interop\Windows\Gdi32\Interop.GetClipRgn.cs" />
     <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.GetCurrentObject.cs"
              Link="Common\Interop\Windows\Gdi32\Interop.GetCurrentObject.cs" />
     <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.GetDeviceCaps.cs"
              Link="Common\Interop\Windows\Gdi32\Interop.GetDeviceCaps.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.GetDIBits.cs"
+             Link="Common\Interop\Windows\Gdi32\Interop.GetDIBits.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.GetObject.cs"
+             Link="Common\Interop\Windows\Gdi32\Interop.GetObject.cs" />
     <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.GetObjectType.cs"
              Link="Common\Interop\Windows\Gdi32\Interop.GetObjectType.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.GetPaletteEntries.cs"
+             Link="Common\Interop\Windows\Gdi32\Interop.GetPaletteEntries.cs" />
     <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.GetRgnBox.cs"
              Link="Common\Interop\Windows\Gdi32\Interop.GetRgnBox.cs" />
     <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.GetStockObject.cs"
              Link="Common\Interop\Windows\Gdi32\Interop.GetStockObject.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.IntersectClipRect.cs"
+             Link="Common\Interop\Windows\Gdi32\Interop.IntersectClipRect.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.ResetDC.cs"
+             Link="Common\Interop\Windows\Gdi32\Interop.ResetDC.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.StartDoc.cs"
+             Link="Common\Interop\Windows\Gdi32\Interop.StartDoc.cs" />
+    <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.StartPage.cs"
+             Link="Common\Interop\Windows\Gdi32\Interop.StartPage.cs" />
     <Compile Include="$(CommonPath)Interop\Windows\Ole32\Interop.IStream.COMWrappers.cs"
              Link="Common\Interop\Windows\Ole32\Interop.IStream.COMWrappers.cs" />
     <Compile Include="$(CommonPath)Interop\Windows\Gdi32\Interop.ObjectType.cs"
